Bring PBE into full PHPCS compliance
