+++ bin/phpcs-whitelist.js
@@ -29,6 +29,8 @@ module.exports = [
 	'modules/memberships/',
 	'modules/module-extras.php',
 	'modules/module-info.php',
+	'modules/post-by-email.php',
+	'modules/post-by-email/',
 	'modules/search/class-jetpack-instant-search.php',
 	'modules/search/class-jetpack-search-customize.php',
 	'modules/search/class.jetpack-search-options.php',
